requirements.txt
Pillow
PySimpleGUI
hentai